refactored award type toggle
