+++ src/js/components/bulkDownload/archive/AwardDataArchiveForm.jsx
@@ -3,14 +3,13 @@
  * Created by Lizzie Salita 12/13/17
  */
 
-import React, { useState, useRef, useEffect } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
 
 import { currentFiscalYear, earliestFiscalYear } from "helpers/fiscalYearHelper";
 import ComboBox from "components/sharedComponents/ComboBox";
-import ArchiveTypeFilter from './filters/TypeFilter';
 import FilterSelectionTitle from "../FilterSelectionTitle";
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
+import AwardTypeToggle from "./filters/AwardTypeToggle";
 
 const currentFY = currentFiscalYear();
 const fyOptions = [];
@@ -72,15 +71,6 @@ const AwardDataArchiveForm = ({
         updateFilter('fy', target.value);
     };
 
-    const onAwardTypeSelect = (e) => {
-        e.preventDefault();
-        const target = e.target;
-        updateFilter('type', {
-            name: target.value,
-            display: target.name
-        });
-    }
-
     return (
         <>
             <FilterSelectionTitle type="agencyFy" />
@@ -99,33 +89,7 @@ const AwardDataArchiveForm = ({
                         placeholder={"Select a Fiscal Year"}
                         defaultValue={`FY ${currentFY}`} />
                 </div>
-                <div className="award-type__container">
-                    View Award Type:
-                    <div className="award-type__buttons">
-                        <button
-                            className={`view-button${
-                                filters.type.name === "contracts" ? " active" : ""
-                            }`}
-                            value={"contracts"}
-                            title={"Contracts"}
-                            aria-label={"Contracts"}
-                            name={"Contracts"}
-                            onClick={onAwardTypeSelect}>
-                            Contracts
-                        </button>
-                        <button
-                            className={`view-button${
-                                filters.type.name === "assistance" ? " active" : ""
-                            }`}
-                            value={"assistance"}
-                            title={"Financial Assistance"}
-                            aria-label={"Financial Assistance"}
-                            name={"Financial Assistance"}
-                            onClick={onAwardTypeSelect}>
-                            Financial Assistance
-                        </button>
-                    </div>
-                </div>
+                <AwardTypeToggle filters={filters} updateFilter={updateFilter} />
             </form>
         </>
     );
